Migrating from sails to Hapi.js
Now using gulp instead of grunt. Hapi is setup and you can hit the login page, but the back-end api still needs to be reworked a bit. For the most part everything has been moved around to the proper spots, but I still need to finish refactoring and fixing things to work with hapi instead of sails.